Changed process of adding new comment, added method for showing info about current user, added exception handling controller.

+++ src/main/java/web/services/CommentService.java
@@ -4,6 +4,9 @@ import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import web.dao.CommentRatingDb;
@@ -37,12 +40,22 @@ public class CommentService {
         session.saveOrUpdate(commentRatingDb);
     }
 
-    public void deleteComment(String id) throws HibernateException {
+    public void deleteComment(String id) throws HibernateException, IndexOutOfBoundsException {
         if (id == null || id.isEmpty()) {
             throw new IllegalArgumentException("Id should not be null or empty");
         }
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if ((authentication instanceof AnonymousAuthenticationToken)) {
+            throw new IllegalArgumentException("User is not logged in");
+        }
+
         Session session = sessionFactory.getCurrentSession();
-        session.createQuery("delete from CommentRatingDb c where c.id=" + id).executeUpdate();
+
+        CommentRatingDb commentToDelete = (CommentRatingDb) session.createQuery("from CommentRatingDb c where c.id=" + id).list().get(0);
+        if (commentToDelete.getClientByIdClient().getLogin().equals(authentication.getName())) {
+            session.createQuery("delete from CommentRatingDb c where c.id=" + id).executeUpdate();
+        }
     }
 
     public CommentRatingDb convertToCommentRatingDB(CommentJSON commentJSON) throws ParsingJsonToDaoException, ParseException {
@@ -50,12 +63,17 @@ public class CommentService {
             throw new IllegalArgumentException("CommentJSON should not be null");
         }
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if ((authentication instanceof AnonymousAuthenticationToken)) {
+            throw new IllegalArgumentException("User is not logged in");
+        }
+
         CommentRatingDb commentRatingDb = new CommentRatingDb();
 
         commentRatingDb.setId(commentJSON.getId());
         commentRatingDb.setRating(commentJSON.getRating());
         commentRatingDb.setFilmByIdFilm(filmService.getFilmWithId(Integer.toString(commentJSON.getIdFilm())));
-        commentRatingDb.setClientByIdClient(clientService.getClientById(Integer.toString(commentJSON.getIdFilm())));
+        commentRatingDb.setClientByIdClient(clientService.getClientByLogin(authentication.getName()));
 
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
         Date parsed = format.parse(commentJSON.getCommentDate());
@@ -67,12 +85,12 @@ public class CommentService {
         return commentRatingDb;
     }
 
-    public List<CommentJSON> getAllCommentsForFilm(String id) throws HibernateException {
+    public List<CommentRatingDb> getAllCommentsForFilm(String id) throws HibernateException {
         if (id == null || id.isEmpty()) {
             throw new IllegalArgumentException("Id should not be null or empty");
         }
         Session session = sessionFactory.getCurrentSession();
-        List<CommentJSON> allComments = session.createQuery("from CommentRatingDb where filmByIdFilm=" + id).list();
+        List<CommentRatingDb> allComments = session.createQuery("from CommentRatingDb where filmByIdFilm=" + id).list();
         return allComments;
     }
